actions/ci: Set loglevel with helm CLI, don't change STUN listener
This makes sure, that the helm upgrade step does not restart the statefulset, as enabling the STUN listeners triggers a change in the statefulset itself (port mapping) which leads to a redeployment of the statefulset.

In contrast, changing the loglevel is only a change in a configmap, hence the sidecar will only trigger a configuration reload of the ejabberd pods.